+++ app/Http/Controllers/Api/SystemController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Helpers\Helper;
+use App\Models\ExamSchedule;
 use App\Repo\Interfaces\BranchInterface;
 use App\Repo\Interfaces\RoomInterface;
 use App\Repo\Interfaces\SystemInterface;
@@ -48,13 +49,18 @@ class SystemController extends Controller
 
     public function deleteSystem( $id){
         try {
-            $response =  $this->system->deleteSystem($id);
-            if($response['status']){
-                $response= Helper::success($response['data'],$response['message']);
-            }else{
-                $response= Helper::error($response['message'],$response['data']);
+            if($exam=ExamSchedule::where('system_id',$id)->count()==0) {
+                $response = $this->system->deleteSystem($id);
+                if ($response['status']) {
+                    $response = Helper::success($response['data'], $response['message']);
+                } else {
+                    $response = Helper::error($response['message'], $response['data']);
+                }
+                return $response;
             }
-            return $response;
+            else{
+                return Helper::error($exam.'exams associated with this room',[]);
+            }
         } catch (\Exception $e) {
             return Helper::error($e->getMessage(),$e);
         }
@@ -68,6 +74,22 @@ class SystemController extends Controller
                 $response= Helper::success($response['data'],$response['message']);
             }else{
                 $response= Helper::error($response['message'],$response['data']);
+            }
+            return $response;
+        } catch (\Exception $e) {
+            return Helper::error($e->getMessage(),$e);
+        }
+    }
+
+    //checkSystemIp
+    public function checkSystemIp($systemIp){
+        try {
+
+            $response =$this->system->checkSystemIp($systemIp);
+            if($response){
+                $response= Helper::success($response,'Ip found');
+            }else{
+                $response= Helper::error('Ip not found',[]);
             }
             return $response;
         } catch (\Exception $e) {
